+++ R/us_counties_daily.R
@@ -1,5 +1,5 @@
 # Functions Roxygen format
-#' @title COVID-19 Counts for US Counties
+#' @title COVID-19 Counts for US Counties from The New York Times
 #'
 #' @description This functions pulls death COVID-19 confirmed and death data
 #'   for US counties from The New York Times GitHub repository downloads it.