+++ NEWS.md
@@ -1,5 +1,11 @@
+# R.COVID.19 0.0.1.2
+* Add new US data from John Hopkins, 2 new functions, which includes geographic
+* data for mapping such as FIPS codes, etc.
+* Add new Global recovery data from John Hopkins, 1 new function.
+* Fixed CRAN re-submission issues.
+
 # R.COVID.19 0.0.1.1
-* Fixed CRAN submission issues
+* Fixed CRAN submission issues.
 * Including function name changes,
 * which are breaking fixes.